+++ ReleaseNotes.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## 2.0.18 - 26.07.2020 - Thomas Kurth
+
+- Bugfix to include App Config assignments
+- Improve Conditional Access Documentation
+- Generate CSV for COnditional Access Documentation
+
+
 ## 2.0.17 - 26.07.2020 - Thomas Kurth
 
 - Bugfix to include MAM assignments